requirements.txt
crewai
streamlit
langchain
langchain-community
langchain-openai
openai
duckduckgo-search
python-dotenv